Added check of var

+++ build.sh
@@ -7,8 +7,13 @@ if [ -z $DOCKER_USERNAME ]; then
   exit 1
 fi
 
+if [ -z $AVD_IMAGE_IP ]; then
+  echo "AVD_IMAGE_IP undefined"
+  exit 1
+fi
+
 rm -f avd.tar.gz
-wget http://155.94.195.190/avd.tar.gz
+wget http://$AVD_IMAGE_IP/avd.tar.gz
 
 docker build --tag $DOCKER_USERNAME/$DOCKER_IMAGE:$DOCKER_TAG \
              --tag $DOCKER_USERNAME/$DOCKER_IMAGE:latest .
